Add exercises up to 4.21*

+++ part4/bloglist/controllers/blogs.js
@@ -71,11 +71,30 @@ blogRouter.put('/:id', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
+    if (!request.token) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
     try {
-        await Blog.findByIdAndDelete(request.params.id);
-        response.status(204).end();
-    } catch(error) {
-        response.status(400).json({ error: error.message });
+       var decodedToken = jwt.verify(request.token, process.env.SECRET);
+    } catch (error) {
+        if (error instanceof jwt.JsonWebTokenError) {
+            return response.status(401).json({ error: 'token missing or invalid' });
+        } else {
+            return response.status(401).json({ error: 'unknown token error' });
+        }
+    }
+
+    blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() === decodedToken.id) {
+        try {
+            await Blog.findByIdAndDelete(request.params.id);
+            response.status(204).end();
+        } catch(error) {
+            response.status(400).json({ error: error.message });
+        }
+    } else {
+        response.status(401).json({ error: 'Unauthorized' });
     }
 });
